move MintInfo to separate file
